+++ packages/repco-frontend/app/routes/$playlists/add.tsx
@@ -39,10 +39,19 @@ export default function Playlists() {
 
   return (
     <div className="px-6 py-6">
+      {added && (
+        <h1 className="font-medium leading-tight text-2xl mt-0 mb-2 text-green-600">
+          Success <br /> {playlistAndUid}
+        </h1>
+      )}
       {playlists.length !== 0 ? (
         playlists.map((e: any) => (
-          <Form key={e} method="post">
-            <li className="font-medium">
+          <Form
+            className="flex-col p-4 w-full justify-center text-center bg-white rounded-lg border shadow-md dark:bg-gray-800 dark:border-gray-700"
+            key={e}
+            method="post"
+          >
+            <p>
               <NavLink
                 className=" text-sm px-0 py-4 font-light text-blue-600 dark:text-blue-500 hover:underline"
                 prefetch="render"
@@ -50,21 +59,14 @@ export default function Playlists() {
               >
                 {e}
               </NavLink>
-              <button
-                className="inline-flex items-center justify-center w-6 h-6 ml-2 text-white transition-colors duration-150 bg-blue-700 rounded-lg focus:shadow-outline hover:bg-blue-800"
-                value={[e, uid]}
-                name="add-to-playlist"
-                type="submit"
-              >
-                <svg className="w-4 h-4 fill-current" viewBox="0 0 20 20">
-                  <path
-                    d="M10 3a1 1 0 011 1v5h5a1 1 0 110 2h-5v5a1 1 0 11-2 0v-5H4a1 1 0 110-2h5V4a1 1 0 011-1z"
-                    clip-rule="evenodd"
-                    fill-rule="evenodd"
-                  ></path>
-                </svg>
-              </button>
-            </li>
+            </p>
+            <button
+              className=" text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm px-4 py-2 dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800"
+              value={[e, uid]}
+              name="add-to-playlist"
+            >
+              add to playlist
+            </button>
           </Form>
         ))
       ) : (
@@ -73,11 +75,6 @@ export default function Playlists() {
             <Link to={'/playlists/new'}>Create at least one Playlist</Link>
           </h1>
         </div>
-      )}
-      {added && (
-        <h1 className="font-medium leading-tight text-2xl mt-0 mb-2 text-green-600">
-          Success <br /> {playlistAndUid}
-        </h1>
       )}
     </div>
   )
